requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.7.2=haa95532_0
ffmpeg-python=0.2.0=pypi_0
future=1.0.0=pypi_0
libffi=3.4.4=hd77b12b_1
openssl=3.0.15=h827c3e9_0
pip=24.2=py310haa95532_0
python=3.10.14=he1021f5_1
setuptools=72.1.0=py310haa95532_0
sqlite=3.45.3=h2bbff1b_0
tk=8.6.14=h0416ee5_0
tzdata=2024a=h04d1e81_0
vc=14.40=h2eaa2aa_1
vs2015_runtime=14.40.33807=h98bb1dd_1
wheel=0.44.0=py310haa95532_0
xz=5.4.6=h8cc25b3_1
zlib=1.2.13=h8cc25b3_1
